+++ Slam.h
@@ -63,7 +63,7 @@ private:
   void setPiName();
 
   /// The recorded times for each image capture.
-  double* captureTimes = NULL;
+  int64_t* captureTimes = NULL;
   /// Whether the captured images have been written to disk or not.
   bool* captureWritten = NULL;
   /// The captured images.
